Move configureStore to src level
